refactor: fix frontend lint warnings
