+++ src/components/Node.tsx
@@ -13,22 +13,8 @@ import Flag from "./Flag";
 import { useTranslation } from "react-i18next";
 import Tips from "./ui/tips";
 
+import { formatUptime } from "@/utils/formatUptime";
 import { formatBytes } from "@/utils/unitHelper";
-
-/** 格式化秒*/
-export function formatUptime(seconds: number, t: TFunction): string {
-  if (!seconds || seconds < 0) return t("nodeCard.time_second", { val: 0 });
-  const d = Math.floor(seconds / 86400);
-  const h = Math.floor((seconds % 86400) / 3600);
-  const m = Math.floor((seconds % 3600) / 60);
-  const s = Math.floor(seconds % 60);
-  const parts = [];
-  if (d) parts.push(`${d} ${t("nodeCard.time_day")}`);
-  if (h) parts.push(`${h} ${t("nodeCard.time_hour")}`);
-  if (m) parts.push(`${m} ${t("nodeCard.time_minute")}`);
-  if (s || parts.length === 0) parts.push(`${s} ${t("nodeCard.time_second")}`);
-  return parts.join(" ");
-}
 
 interface NodeProps {
   basic: NodeBasicInfo;
@@ -295,7 +281,6 @@ type NodeGridProps = {
 };
 
 import { Box } from "@radix-ui/themes";
-import type { TFunction } from "i18next";
 import { Link } from "react-router-dom";
 import { useIsMobile } from "@/hooks/use-mobile";
 import type { NodeBasicInfo } from "@/contexts/NodeListContext";
