+++ src/contexts/NodeListContext.ts
@@ -1,0 +1,60 @@
+import React from "react";
+
+export type NodeBasicInfo = {
+  uuid: string;
+  name: string;
+  cpu_name: string;
+  virtualization: string;
+  arch: string;
+  cpu_cores: number;
+  os: string;
+  kernel_version: string;
+  gpu_name: string;
+  region: string;
+  mem_total: number;
+  swap_total: number;
+  disk_total: number;
+  version: string;
+  weight: number;
+  price: number;
+  tags: string;
+  billing_cycle: number;
+  currency: string;
+  group: string;
+  traffic_limit: number;
+  traffic_limit_type: undefined | "sum" | "max" | "min" | "up" | "down";
+  expired_at: string;
+  created_at: string;
+  updated_at: string;
+  ipv4?: string;
+  ipv6?: string;
+};
+
+export interface NodeListContextType {
+  nodeList: NodeBasicInfo[] | null;
+  isLoading: boolean;
+  error: string | null;
+  refresh: () => void;
+}
+
+const NODE_LIST_CONTEXT_KEY = "__komariNodeListContext" as const;
+
+type NodeListContextGlobal = typeof globalThis & {
+  [NODE_LIST_CONTEXT_KEY]?: React.Context<NodeListContextType | undefined>;
+};
+
+const globalNodeListContext = globalThis as NodeListContextGlobal;
+export const NodeListContext =
+  globalNodeListContext[NODE_LIST_CONTEXT_KEY] ??
+  (globalNodeListContext[NODE_LIST_CONTEXT_KEY] =
+    React.createContext<NodeListContextType | undefined>(undefined));
+
+export function useNodeList(): NodeListContextType;
+export function useNodeList(required: false): NodeListContextType | undefined;
+export function useNodeList(required = true) {
+  const context = React.useContext(NodeListContext);
+  if (!context && required) {
+    throw new Error("useNodeList must be used within a NodeListProvider");
+  }
+  return context;
+}
