+++ src/contexts/PublicInfoContext.ts
@@ -1,0 +1,43 @@
+import React from "react";
+
+export interface PublicInfo {
+  cors_origin_check_enabled: boolean;
+  custom_body: string;
+  custom_head: string;
+  description: string;
+  disable_password_login: boolean;
+  oauth_provider: string;
+  oauth_enable: boolean;
+  metric_retention_days: number;
+  sitename: string;
+  private_site: boolean;
+  theme: string;
+  theme_settings: any;
+  [property: string]: any;
+}
+
+export interface PublicInfoResponse {
+  data: PublicInfo;
+  message: string;
+  status: string;
+  [property: string]: any;
+}
+
+export interface PublicInfoContextType {
+  publicInfo: PublicInfo | null;
+  isLoading: boolean;
+  error: string | null;
+  refresh: () => Promise<void>;
+}
+
+export const PublicInfoContext = React.createContext<
+  PublicInfoContextType | undefined
+>(undefined);
+
+export const usePublicInfo = () => {
+  const context = React.useContext(PublicInfoContext);
+  if (!context) {
+    throw new Error("usePublicInfo must be used within a PublicInfoProvider");
+  }
+  return context;
+};
