+++ src/contexts/RPC2Provider.tsx
@@ -1,25 +1,12 @@
 import React, {
-  createContext,
   useCallback,
-  useContext,
   useEffect,
   useMemo,
   useState,
 } from "react";
-import { RPC2Client } from "../lib/rpc2";
-import type { RPC2ConnectionStateType } from "../types/rpc2";
 import i18n from "../i18n/config";
-
-interface RPC2ContextType {
-  client: RPC2Client;
-  connectionState: RPC2ConnectionStateType;
-  isConnected: boolean;
-  error: string | null;
-  connect: () => Promise<void>;
-  disconnect: () => void;
-}
-
-const RPC2Context = createContext<RPC2ContextType | undefined>(undefined);
+import { RPC2Client } from "../lib/rpc2";
+import { RPC2Context } from "./RPC2Context";
 
 // 模块级单例，避免在开发环境 StrictMode 或路由切换时产生多个连接
 let __rpc2_singleton__: RPC2Client | null = null;
@@ -101,46 +88,3 @@ export const RPC2Provider: React.FC<{ children: React.ReactNode }> = ({ children
     </RPC2Context.Provider>
   );
 };
-
-export const useRPC2 = (): RPC2ContextType => {
-  const context = useContext(RPC2Context);
-  if (context === undefined) {
-    throw new Error(i18n.t("rpc2.provider_required"));
-  }
-  return context;
-};
-
-// 自定义 Hook 用于调用 RPC 方法
-export const useRPC2Call = () => {
-  const { client, isConnected } = useRPC2();
-
-  // 保持稳定引用，避免消费者重复触发副作用
-  const call = useCallback(<TParams = any, TResult = any>(
-    method: string,
-    params?: TParams,
-    options?: any
-  ): Promise<TResult> => client.call(method, params, options), [client]);
-
-  const callViaWebSocket = useCallback(<TParams = any, TResult = any>(
-    method: string,
-    params?: TParams,
-    options?: any
-  ): Promise<TResult> => client.callViaWebSocket(method, params, options), [client]);
-
-  const callViaHTTP = useCallback(<TParams = any, TResult = any>(
-    method: string,
-    params?: TParams,
-    options?: any
-  ): Promise<TResult> => client.callViaHTTP(method, params, options), [client]);
-
-  const batchCall = useCallback((requests: Array<{ method: string; params?: any; notification?: boolean }>) =>
-    client.batchCall(requests), [client]);
-
-  return {
-    call,
-    callViaWebSocket,
-    callViaHTTP,
-    batchCall,
-    isConnected,
-  };
-}
