+++ src/pages/admin/_layout.tsx
@@ -1,8 +1,8 @@
 import { Outlet } from "react-router-dom";
 
 import AdminPanelBar from "../../components/admin/AdminPanelBar";
-import { AdminNavigationProvider } from "@/contexts/AdminNavigationContext";
-import { AccountProvider } from "@/contexts/AccountContext";
+import { AdminNavigationProvider } from "@/contexts/AdminNavigationProvider";
+import { AccountProvider } from "@/contexts/AccountProvider";
 import { updateSettingsWithToast, useSettings } from "@/lib/api";
 import { Button, Dialog } from "@radix-ui/themes";
 import { useEffect, useState } from "react";
