+++ src/pages/admin/account.tsx
@@ -2,7 +2,8 @@ import React from "react";
 
 import { toast } from "sonner";
 import { useTranslation } from "react-i18next";
-import { AccountProvider, useAccount } from "@/contexts/AccountContext";
+import { useAccount } from "@/contexts/AccountContext";
+import { AccountProvider } from "@/contexts/AccountProvider";
 import {
   Badge,
   Button,
@@ -377,7 +378,7 @@ const TwoFactorDisabled = () => {
         .catch((err) => toast.error(err.message))
         .finally(() => setIsLoading(false));
     }
-  }, [isOpen]);
+  }, [isOpen, t]);
 
   const handleEnable2fa = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
