+++ src/pages/admin/notification/load.tsx
@@ -9,14 +9,12 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import {
-  LoadAlertProvider,
   useLoadAlert,
   type LoadAlert,
 } from "@/contexts/LoadAlertContext";
-import {
-  NodeDetailsProvider,
-  useNodeDetails,
-} from "@/contexts/NodeDetailsContext";
+import { LoadAlertProvider } from "@/contexts/LoadAlertProvider";
+import { useNodeDetails } from "@/contexts/NodeDetailsContext";
+import { NodeDetailsProvider } from "@/contexts/NodeDetailsProvider";
 
 import {
   Button,
