+++ src/pages/admin/themes.tsx
@@ -139,7 +139,7 @@ const ThemePage = () => {
 
   const loading = themesLoading || settingsLoading || !currentTheme;
   // 获取主题列表
-  const fetchThemes = async () => {
+  const fetchThemes = useCallback(async () => {
     try {
       const response = await fetch("/api/admin/theme/list");
       if (!response.ok) {
@@ -160,7 +160,7 @@ const ThemePage = () => {
     } finally {
       setThemesLoading(false);
     }
-  };
+  }, [currentTheme]);
 
   // 上传主题
   const uploadTheme = async (file: File) => {
@@ -369,7 +369,7 @@ const ThemePage = () => {
   // 同步活跃状态
   useEffect(() => {
     fetchThemes();
-  }, [currentTheme]);
+  }, [fetchThemes]);
 
   useEffect(() => {
     if (!settingsLoading && themes.length > 0) {
