+++ src/pages/terminal/CommandClipboard.tsx
@@ -1,10 +1,10 @@
 import LanguageSwitch from "@/components/Language";
 import Loading from "@/components/loading";
 import {
-  CommandClipboardProvider,
   useCommandClipboard,
   type CommandClipboard,
 } from "@/contexts/CommandClipboardContext";
+import { CommandClipboardProvider } from "@/contexts/CommandClipboardProvider";
 import { useTerminal } from "@/contexts/TerminalContext";
 import {
   Button,
